Return DTO or an array by mapper instead of entity directly
