+++ src/SalaryHistory/Application/DTOs/StoreSalaryHistoryDTO.php
@@ -3,7 +3,6 @@
 namespace Src\SalaryHistory\Application\DTOs;
 
 use Illuminate\Http\Request;
-use Illuminate\Support\Collection;
 use Src\SalaryHistory\Domain\ValueObjects\Currency;
 use Src\SalaryHistory\Domain\ValueObjects\Salary;
 use Src\Shared\Domain\ValueObjects\Date;
@@ -57,27 +56,6 @@ final class StoreSalaryHistoryDTO
             currency: Currency::fromString($request->currency),
             note: $request->note
         );
-    }
-
-    /**
-     * @param mixed $entities SalaryHistory Entities
-     * 
-     * @return array
-     */
-    public static function toResponse(mixed $entities): array
-    {
-        $collection = $entities instanceof Collection ? $entities : collect([$entities]);
-       
-        return $collection->map(function ($entity) {
-            return [
-                'id' => $entity->getId(),
-                'user_id' => $entity->getUserId(),
-                'on_date' => $entity->getOnDate()->format(),
-                'salary' => $entity->getSalary()->toString(),
-                'currency' => $entity->getCurrency()->toString(),
-                'note' => $entity->getNote(),
-            ];
-        })->toArray();
     }
 
     /**
