+++ src/SalaryHistory/Application/Mappers/SalaryHistoryMapper.php
@@ -1,6 +1,7 @@
 <?php
 namespace Src\SalaryHistory\Application\Mappers;
 
+use Illuminate\Support\Collection;
 use Src\Shared\Domain\Exceptions\FactoryException;
 use Src\Shared\Domain\ValueObjects\Date;
 use Src\SalaryHistory\Application\DTOs\StoreSalaryHistoryDTO;
@@ -56,5 +57,43 @@ class SalaryHistoryMapper
             
             throw new FactoryException('Error creating SalaryHistory by Mapper: ' . $e->getMessage());
         }
+    }
+
+    /**
+     * @param mixed $entities
+     * 
+     * @return array
+     */
+    public static function toResponse(mixed $entities): array
+    {
+        $collection = $entities instanceof Collection ? $entities : collect([$entities]);
+       
+        return $collection->map(function (SalaryHistory $entity) {
+            return [
+                'id' => $entity->getId(),
+                'user_id' => $entity->getUserId(),
+                'on_date' => $entity->getOnDate()->format(),
+                'salary' => $entity->getSalary()->toString(),
+                'currency' => $entity->getCurrency()->toString(),
+                'note' => $entity->getNote(),
+            ];
+        })->toArray();
+    }
+
+    /**
+     * @param SalaryHistory $salaryHistory
+     * 
+     * @return array
+     */
+    public static function toArray(SalaryHistory $salaryHistory): array
+    {
+        return [
+            'id' => $salaryHistory->getId(),
+            'user_id' => $salaryHistory->getUserId(),
+            'on_date' => $salaryHistory->getOnDate()->format(),
+            'salary' => $salaryHistory->getSalary()->getAmount(),
+            'currency' => $salaryHistory->getCurrency()->toString(),
+            'note' => $salaryHistory->getNote()
+        ];
     }
 }
